# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: contact@josehernandez.dev
    name: Smidgen
    url: https://example.com/
  description: API for interacting with Smidgen.
  license:
    name: " GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007"
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  title: Smidgen
  version: 0.0.1
servers:
- url: /api/v1
paths:
...
  /equipment_assignment/:
    get:
      description: Get all assignments stored in the database.
      operationId: get_assignment_equipment_assignment__get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/equipment_assignment'
                title: Response Get Assignment Equipment Assignment  Get
                type: array
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
      summary: Get assignments
      tags:
      - equipment assignment
    post:
      description: Create a new assignment in the database.
      operationId: add_assignment_equipment_assignment__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/equipment_assignment'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create assignment
      tags:
      - equipment assignment
  /equipment_assignment/{assignment_id}:
    delete:
      description: Delete the specified assignment from the database.
      operationId: delete_assignment_equipment_assignment__assignment_id__delete
      parameters:
      - explode: false
        in: path
        name: assignment_id
        required: true
        schema:
          title: Assignment Id
          type: integer
        style: simple
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete assignment
      tags:
      - equipment assignment
    get:
      description: Get the specified assignment stored in the database.
      operationId: get_assignments_equipment_assignment__assignment_id__get
      parameters:
      - explode: false
        in: path
        name: assignment_id
        required: true
        schema:
          title: Assignment Id
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/equipment_assignment'
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get assignment
      tags:
      - equipment assignment
    put:
      description: Update the specified assignment in the database.
      operationId: update_assignment_equipment_assignment__assignment_id__put
      parameters:
      - explode: false
        in: path
        name: assignment_id
        required: true
        schema:
          title: Assignment Id
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/equipment_assignment'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update assignment
      tags:
      - equipment assignment
...
components:
  schemas:
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          type: type
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          type: type
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      example:
        msg: msg
        loc:
        - Location_inner
        - Location_inner
        type: type
      properties:
        loc:
          items:
            $ref: '#/components/schemas/Location_inner'
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    equipment_assignment:
      example:
        assignment_id: 0
        user_id: 6
        date_of_assignment: 2000-01-23T04:56:07.000+00:00
        equipment_id: 1
      properties:
        assignment_id:
          title: Assignment Id
          type: integer
        user_id:
          title: User Id
          type: integer
        equipment_id:
          title: Equipment Id
          type: integer
        date_of_assignment:
          format: date-time
          title: Date Of Assignment
          type: string
      required:
      - assignment_id
      - date_of_assignment
      - equipment_id
      - user_id
      title: equipment_assignment
      type: object
...
    Location_inner:
      anyOf:
      - type: string
      - type: integer