# go/api/equipment_assignment.go
/*
 * api
 *
 * API for interacting with api.
 *
 * API version: 0.0.1
 * Contact: contact@josehernandez.dev
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package smidgen

import (
	"encoding/json"
	"net/http"
	models "smidgen-backend/go/models"
	utils "smidgen-backend/go/utils"
	"strings"

	"github.com/gorilla/mux"
)

// EquipmentAssignmentAPIController binds http requests to an api service and writes the service results to the http response
type EquipmentAssignmentAPIController struct {
	service      EquipmentAssignmentAPIServicer
	errorHandler utils.ErrorHandler
}

// EquipmentAssignmentAPIOption for how the controller is set up.
type EquipmentAssignmentAPIOption func(*EquipmentAssignmentAPIController)

// WithEquipmentAssignmentAPIErrorHandler inject ErrorHandler into controller
func WithEquipmentAssignmentAPIErrorHandler(h utils.ErrorHandler) EquipmentAssignmentAPIOption {
	return func(c *EquipmentAssignmentAPIController) {
		c.errorHandler = h
	}
}

// NewEquipmentAssignmentAPIController creates a default api controller
func NewEquipmentAssignmentAPIController(s EquipmentAssignmentAPIServicer, opts ...EquipmentAssignmentAPIOption) utils.Router {
	controller := &EquipmentAssignmentAPIController{
		service:      s,
		errorHandler: utils.DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the EquipmentAssignmentAPIController
func (c *EquipmentAssignmentAPIController) Routes() utils.Routes {
	return utils.Routes{
		"AddAssignmentEquipmentAssignmentPost": utils.Route{
			Method:      strings.ToUpper("Post"),
			Pattern:     "equipment_assignment/",
			HandlerFunc: c.AddAssignmentEquipmentAssignmentPost,
		},
		"DeleteAssignmentEquipmentAssignmentAssignmentIdDelete": utils.Route{
			Method:      strings.ToUpper("Delete"),
			Pattern:     "equipment_assignment/{assignment_id}",
			HandlerFunc: c.DeleteAssignmentEquipmentAssignmentAssignmentIdDelete,
		},
		"GetAssignmentEquipmentAssignmentGet": utils.Route{
			Method:      strings.ToUpper("Get"),
			Pattern:     "equipment_assignment/",
			HandlerFunc: c.GetAssignmentEquipmentAssignmentGet,
		},
		"GetAssignmentsEquipmentAssignmentAssignmentIdGet": utils.Route{
			Method:      strings.ToUpper("Get"),
			Pattern:     "equipment_assignment/{assignment_id}",
			HandlerFunc: c.GetAssignmentsEquipmentAssignmentAssignmentIdGet,
		},
		"UpdateAssignmentEquipmentAssignmentAssignmentIdPut": utils.Route{
			Method:      strings.ToUpper("Put"),
			Pattern:     "equipment_assignment/{assignment_id}",
			HandlerFunc: c.UpdateAssignmentEquipmentAssignmentAssignmentIdPut,
		},
	}
}

// AddAssignmentEquipmentAssignmentPost - Create assignment
func (c *EquipmentAssignmentAPIController) AddAssignmentEquipmentAssignmentPost(w http.ResponseWriter, r *http.Request) {
	equipmentAssignmentParam := models.EquipmentAssignment{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&equipmentAssignmentParam); err != nil {
		c.errorHandler(w, r, &utils.ParsingError{Err: err}, nil)
		return
	}
	if err := models.AssertEquipmentAssignmentRequired(equipmentAssignmentParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := models.AssertEquipmentAssignmentConstraints(equipmentAssignmentParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.AddAssignmentEquipmentAssignmentPost(r.Context(), equipmentAssignmentParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	utils.EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteAssignmentEquipmentAssignmentAssignmentIdDelete - Delete assignment
func (c *EquipmentAssignmentAPIController) DeleteAssignmentEquipmentAssignmentAssignmentIdDelete(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	assignmentIdParam, err := utils.ParseNumericParameter[int32](
		params["assignment_id"],
		utils.WithRequire[int32](utils.ParseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &utils.ParsingError{Err: err}, nil)
		return
	}
	result, err := c.service.DeleteAssignmentEquipmentAssignmentAssignmentIdDelete(r.Context(), assignmentIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	utils.EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetAssignmentEquipmentAssignmentGet - Get assignments
func (c *EquipmentAssignmentAPIController) GetAssignmentEquipmentAssignmentGet(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetAssignmentEquipmentAssignmentGet(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	utils.EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetAssignmentsEquipmentAssignmentAssignmentIdGet - Get assignment
func (c *EquipmentAssignmentAPIController) GetAssignmentsEquipmentAssignmentAssignmentIdGet(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	assignmentIdParam, err := utils.ParseNumericParameter[int32](
		params["assignment_id"],
		utils.WithRequire[int32](utils.ParseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &utils.ParsingError{Err: err}, nil)
		return
	}
	result, err := c.service.GetAssignmentsEquipmentAssignmentAssignmentIdGet(r.Context(), assignmentIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	utils.EncodeJSONResponse(result.Body, &result.Code, w)
}

// UpdateAssignmentEquipmentAssignmentAssignmentIdPut - Update assignment
func (c *EquipmentAssignmentAPIController) UpdateAssignmentEquipmentAssignmentAssignmentIdPut(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	assignmentIdParam, err := utils.ParseNumericParameter[int32](
		params["assignment_id"],
		utils.WithRequire[int32](utils.ParseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &utils.ParsingError{Err: err}, nil)
		return
	}
	equipmentAssignmentParam := models.EquipmentAssignment{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&equipmentAssignmentParam); err != nil {
		c.errorHandler(w, r, &utils.ParsingError{Err: err}, nil)
		return
	}
	if err := models.AssertEquipmentAssignmentRequired(equipmentAssignmentParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := models.AssertEquipmentAssignmentConstraints(equipmentAssignmentParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.UpdateAssignmentEquipmentAssignmentAssignmentIdPut(r.Context(), assignmentIdParam, equipmentAssignmentParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	utils.EncodeJSONResponse(result.Body, &result.Code, w)
}